# Improve performance of DateTime.UtcNow on Windows