Follow-up to https://github.com/dotnet/runtime/pull/44771. Builds atop work done in https://github.com/dotnet/runtime/pull/45479 and https://github.com/dotnet/runtime/pull/46690.
Resolves https://github.com/dotnet/runtime/issues/13091.

This PR is based on a series of conversations @tarekgh and I have had recently with regard to short-circuiting our leap second handling logic where possible. For background, Windows understands that days normally have 86,400 seconds, but this can be off by ±1 second due to positive or negative leap seconds. .NET's DateTime type requires each day to have _exactly_ 86,400 seconds. Previous .NET releases resolve this discrepancy through the following special-cased logic: "Is the current time 23:59:60.xxx? If so, I'll return a DateTime instance whose time component reads 23:59:59.999." Unfortunately, this check is not free, and it shows up in benchmarks when we access `DateTime.UtcNow`.

This PR relies on the following two properties of leap seconds:

* leap seconds only ever kick in on the last second of the day, _and_
* any given day will never be adjusted beyond ±1 second due to leap seconds.

This optimization works by keeping a cache which maps the last-seen Win32 `FILETIME` struct to its equivalent .NET `DateTime` struct. The cache is valid for some window of time (by default, 5 minutes). If a call to `DateTime.UtcNow` takes place during the cache's validity window, we don't bother asking the OS to deconstruct the `FILETIME` struct back into a `SYSTEMTIME` struct (which accounts for leap seconds). Instead, we know that we can just perform simple addition without worrying about leap second handling.

There is one condition we need to check: Does a leap second _actually_ occur within the next 5 minutes? If so, we can't store _(now, now + 5 min)_ into the cache, otherwise `DateTime.UtcNow` will start returning incorrect values as soon as midnight rolls around. To account for this, we slide the cache window backward, choosing to set the cache validity period to _(today at 23:54:59, today at 23:59:59)_ instead of _(now, now + 5 min)_.

The reason we use a 5 minute validity window instead of a 24-hour validity window is called out in code comments. tl;dr: The OS uses shared memory mapping to publish leap second tables to all processes, and it can extend that table without killing active processes or rebooting the machine. There is a slim but non-zero chance that this table can be updated to insert leap second data _into the past_. If this happens, our cache will be incorrect. After consultation with the Windows team, we decided that a 5-minute validity window is an appropriate tradeoff that improves performance in the common case and minimizes the window in which an application might encounter problems if this situation were to occur.

|    Method |        Job | Toolchain |     Mean |    Error |   StdDev | Ratio |
|---------- |----------- |---------- |---------:|---------:|---------:|------:|
| GetUtcNow | Job-JYYDLR |      main | 65.15 ns | 0.103 ns | 0.097 ns |  1.00 |
| GetUtcNow | Job-RHLQLD |     proto | 27.74 ns | 0.359 ns | 0.336 ns |  0.43 |