@GrabYourPitchforks 2021-03-25
Test cases outstanding:

- [ ] Ensure `DateTime.UtcNow` uses cache when inside validity period (line 30)
- [ ] Ensure `DateTime.UtcNow` doesn't use cache when outside validity period (line 33)
- [ ] Ensure cache early-exit is hit if seconds field reads __60_) (line 215)
- [ ] Ensure validity window is shifted back if a positive leap second would occur and that entire validity window is valid (line 253)
- [ ] Ensure validity window is shifted back if a negative leap second would occur and that entire validity window is valid (line 253)