@mattjohnsonpint 2021-03-26
In real life, leap seconds can only occur on two specific dates, June 30th and Dec 31st (UTC).  Would it be worth optimizing further for that?  Or do we want to accept any leap second Windows throws at us?