@GrabYourPitchforks 2021-03-26
@mattjohnsonpint Good question. I don't see any good opportunity for us to take advantage of that. Even in the original abandoned PR (https://github.com/dotnet/runtime/pull/44771), which was pretty aggressive about caching, we didn't look at the date component. We were actually asked to be _less_ aggressive. :) So I knocked the cache validity window down from 23h59m59s to 5 min.

Given this, do you see an opportunity for us to use the date component to increase the size of the window? Nothing comes to mind on my end.