@GrabYourPitchforks 2021-03-26
@tarekgh You mean 4.8? Sure, I can do that. Any other runtimes of interest?

Open question: is it possible that DateTime.UtcNow might now be recursive? Like, some debugging or introspection engine hooking allocations and and calling DateTime.UtcNow as part of its logging mechanism? If this is a possibility, I can add a guard against it.