@tarekgh 2021-03-26
>  You mean 4.8? Sure, I can do that. Any other runtimes of interest?

No, I meant try the tool with your changes. the tool can allow you insert leap seconds. so, you can try run your test and in the middle run the tool to insert a leap second and look if the test continues behave nicely :-) 

> Open question: is it possible that DateTime.UtcNow might now be recursive? Like, some debugging or introspection engine hooking allocations and and calling DateTime.UtcNow as part of its logging mechanism? If this is a possibility, I can add a guard against it.

I am not aware of any case like that which can cause the recursion. I am not even sure if this is possible with UtcNow.