@tarekgh 2021-03-26
> Makes me think we should have a guard against recursive calls, 

Were you able to get into this case? could you please share more info how this can happen?