@GrabYourPitchforks 2021-03-26
> Were you able to get into this case? could you please share more info how this can happen?

I didn't try it. It was just a thought experiment. But in theory:

```cs
try
{
    DoWork();
}
catch when (MyExceptionFilter())
{
    /* blah blah */
}
```

If _DoWork_ throws an exception, _MyExceptionFilter_ could end up running within _DoWork_'s frame. (.NET doesn't unwind the call stack.)

If _DoWork_ calls `DateTime.UtcNow`, and if `DateTime.UtcNow` throws an exception (perhaps an OOM due to the allocation in the cache miss path), and if _MyExceptionFilter_ also calls `DateTime.UtcNow` (perhaps as part of logging), then we're in a recursive call scenario. I really don't think this is likely in practice, but I can't prove to myself that it will _never_ happen.