@GrabYourPitchforks 2021-03-26
@tarekgh I tried all kinds of wonderful and horrible combinations, such as using `w32tm` to add leap seconds while this method was executing, changing the system time while this method was executing, changing the process's "60-second leap seconds" support while this method was executing, and so on. (I didn't try _removing_ leap seconds while this method was executing, as that would violate OS invariants. The `w32tm` tool doesn't allow this without a system reboot anyway.)

In all cases, the cutoff point is the code below.

https://github.com/dotnet/runtime/blob/941e477616cf974269d9f8558ba8863271f75d0f/src/libraries/System.Private.CoreLib/src/System/DateTime.Windows.cs#L223-L228

If the clock / leap seconds are changed _before_ this point, we correctly respect it. Once this point is hit, we're already down the "start calculating values to cache" code path. So if the clock / leap seconds change _after_ this point, we'll end up caching incorrect values for 5 minutes. Then after that 5 minute window is up, we'll evict our cache and query the OS for the actual time again.

To the application, the observed behavior is that the return value of `DateTime.UtcNow` might jump a little bit forward or backward once the cache expires. It seems to be no different than today's behavior if I were to change the system clock by a second or two while the app is running. I didn't see any ill effect.