@tarekgh 2021-03-26
> If DoWork calls DateTime.UtcNow, and if DateTime.UtcNow throws an exception (perhaps an OOM due to the allocation in the cache miss path), and if MyExceptionFilter also calls DateTime.UtcNow (perhaps as part of logging), then we're in a recursive call scenario. I really don't think this is likely in practice, but I can't prove to myself that it will never happen.

I don't think this is a specific problem for that case but it is generic issue which can happen in many ways. 

CC @jkotas if he has any thoughts around that.