@mnmr 2021-03-26
I was wondering if there has been a discussion about changing the way leap seconds are handled? 

Reading this discussion makes me think that the same code running on Windows and Linux will report different times, which is unfortunate. Google has in interesting solution where leap seconds are "distributed" over an hour, which has the nice side effect that time only ever moves forward. Is something like this being considered at all?

Apologies if this is not the right forum for voicing such an idea.