@GrabYourPitchforks 2021-03-26
> Google has in interesting solution where leap seconds are "distributed" over an hour, which has the nice side effect that time only ever moves forward. Is something like this being considered at all?

This is also the default behavior for Windows applications. Applications must opt in to leap second support. If your application has not opted in, the OS fudges the time it reports back to you by slowing down the clock over the last 2 seconds.

Examples:

| Actual time  | GetSystemTime returns |
|--------------|-----------------------|
| 23:59:59.000 | 23:59:59.000          |
| 23:59:59.500 | 23:59:59.250          |
| 23:59:60.000 | 23:59:59.500          |
| 23:59:60.500 | 23:59:59.750          |

The logic proposed by this PR handles this scenario fine.