@jkotas 2021-03-26
> I don't think this is a specific problem for that case but it is generic issue which can happen in many ways.

Yes, OOM can happen at atribitrary points, e.g. due to JITing.

> Makes me think we should have a guard against recursive calls, just so that we don't bring down the app in these weird edge cases

Not worth it. The app is going down once you hit this situation anyway.